Add lang-map package to get language
